MWPW-121113 refactoring after review
